Replaced all the structure asserts with method
we're running this automated so that makes the asserts
bad news. I left the assert for now until I come up with
a better reporting method for malformed QTI.
(Since so many vendors get it wrong anyway)

I made it so that the new method doesn't have to assert if it's
a property we don't really care is out of place. (like random
vendor-specific properties)